# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: developers@givechariot.com
    name: Chariot Development Team
    url: https://www.givechariot.com/contact
  description: The Chariot Payments REST API.
  termsOfService: https://www.givechariot.com/legal-nonprofit
  title: Chariot Payments API
  version: v1
servers:
- description: Staging
  url: https://devapi.givechariot.com
- description: Sandbox
  url: https://sandboxapi.givechariot.com
- description: Production
  url: https://api.givechariot.com
security:
- bearerAuth: []
paths:
  /accounts:
    get:
      description: |
        List accounts.
        This endpoint is used to list accounts.
      operationId: listAccounts
      parameters:
      - description: Limit the size of the list that is returned. The default (and
          maximum) is 100 objects.
        explode: true
        in: query
        name: limit
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - description: "The cursor to use for pagination. If not set, the first page\
          \ of results will be returned."
        explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountList'
          description: The accounts were retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: List accounts
      tags:
      - accounts
    post:
      description: |
        Create a bank account with Chariot.
        This endpoint is used to create a bank account for a payer in the Chariot Giving Payment Network.
        The account is setup in the "For-Benefit-Of" (FBO) model where the account is owned and maintained by Chariot.
        An account is necessary in order to make and send payments to nonprofits.
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
        description: The account to create
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: The account was created
          headers:
            Location:
              $ref: '#/components/headers/Location'
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Create an account
      tags:
      - accounts
  /accounts/{id}:
    get:
      description: |
        Get an account by its unique identifier.
        This endpoint is used to retrieve an account by its unique identifier.
      operationId: getAccount
      parameters:
      - description: The unique identifier for the account
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
          description: The account was retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "404":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 404
                error: Not Found
                message: Resource is not found
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "409":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 409
                error: Conflict
                message: Request conflicts with an already existing resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Conflicts
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Retrieve an account
      tags:
      - accounts
...
  /accounts/{id}/details:
    get:
      description: |
        Get the details for an account.
        This endpoint is used to retrieve the details for a single account.
      operationId: getAccountDetails
      parameters:
      - description: The unique identifier for the account
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountDetails'
          description: The details were retrieved
        "400":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 400
                error: Bad Request
                message: Expected an API key to be provided in the header `x-chariot-api-key`
              schema:
                $ref: '#/components/schemas/ApiError'
          description: The request is invalid or contains invalid parameters
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "401":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 401
                error: Unauthorized
                message: Unauthorized
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Unauthorized. The request is missing the security (OAuth2 Bearer
            token) requirements and the server is unable to verify the identify of
            the caller.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "403":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 403
                error: Forbidden
                message: User is not allowed to access this resource
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Access denied
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
        "404":
          content:
            application/json:
              example:
                timestamp: 2020-07-10 15:00:00.000
                code: 404
                error: Not Found
                message: Resource is not found
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Resource Not Found
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
      summary: Retrieve account details
      tags:
      - accounts
...
components:
  headers:
    X-Request-Id:
      description: The unique identifier for the request
      explode: false
      schema:
        type: string
      style: simple
...
    Location:
      description: The URI reference of the created object.
      explode: false
      schema:
        format: uri
        type: string
      style: simple
...
  schemas:
    Account:
      example:
        us_bank_account_id: "1234567890"
        name: Payer FBO Account
        created_at: 2000-01-23T04:56:07.000+00:00
        id: "1234567890"
        us_bank_account_number_id: "1234567890"
      properties:
        id:
          description: The unique identifier for the account
          example: "1234567890"
          readOnly: true
          type: string
        name:
          description: The name of the account
          example: Payer FBO Account
          type: string
        us_bank_account_id:
          description: The unique identifier for the US bank account
          example: "1234567890"
          readOnly: true
          type: string
        us_bank_account_number_id:
          description: The unique identifier for the US bank account number
          example: "1234567890"
          readOnly: true
          type: string
        created_at:
          description: The date and time the account was created
          format: date-time
          readOnly: true
          type: string
      type: object
    AccountList:
      description: A list of accounts
      example:
        data:
        - us_bank_account_id: "1234567890"
          name: Payer FBO Account
          created_at: 2000-01-23T04:56:07.000+00:00
          id: "1234567890"
          us_bank_account_number_id: "1234567890"
        - us_bank_account_id: "1234567890"
          name: Payer FBO Account
          created_at: 2000-01-23T04:56:07.000+00:00
          id: "1234567890"
          us_bank_account_number_id: "1234567890"
        paging:
          cursors:
            before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
            after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          total: 100
      properties:
        data:
          description: The list of accounts
          items:
            $ref: '#/components/schemas/Account'
          type: array
        paging:
          $ref: '#/components/schemas/Pagination'
      type: object
    AchNumber:
      description: The identifying number used for making electronic transfers to
        an account
      example:
        account_number: "1234567890"
        routing_number: "123456789"
      properties:
        account_number:
          description: The account number
          example: "1234567890"
          readOnly: true
          type: string
        routing_number:
          description: The American Bankers' Association (ABA) Routing Transit Number
            (RTN).
          example: "123456789"
          readOnly: true
          type: string
      type: object
    AccountNumbers:
      description: The identifying numbers used for making electronic transfers to
        accounts
      example:
        ach:
        - account_number: "1234567890"
          routing_number: "123456789"
        - account_number: "1234567890"
          routing_number: "123456789"
      properties:
        ach:
          description: The ACH numbers
          items:
            $ref: '#/components/schemas/AchNumber'
          type: array
      type: object
    AccountDetails:
      description: The details of an account
      example:
        numbers:
          ach:
          - account_number: "1234567890"
            routing_number: "123456789"
          - account_number: "1234567890"
            routing_number: "123456789"
        status: open
      properties:
        numbers:
          $ref: '#/components/schemas/AccountNumbers'
        status:
          description: The status of the account
          enum:
          - open
          - closed
          example: open
          readOnly: true
          type: string
      type: object
...
    Pagination:
      description: The paging information
      example:
        cursors:
          before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
        total: 100
      properties:
        cursors:
          $ref: '#/components/schemas/Pagination_cursors'
        total:
          description: The total number of objects
          example: 100
          type: integer
...
    ApiError:
      description: An API error
      properties:
        timestamp:
          description: The date and time the error occurred
          format: date-time
          readOnly: true
          type: string
        status:
          description: The HTTP status code
          readOnly: true
          type: integer
        error:
          description: The HTTP status message
          readOnly: true
          type: string
        message:
          description: The error message
          readOnly: true
          type: string
      type: object
...
    Pagination_cursors:
      description: The cursors for the next and previous pages
      example:
        before: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
        after: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
      properties:
        before:
          description: The cursor for the previous page
          example: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          type: string
        after:
          description: The cursor for the next page
          example: eyJpZCI6IjEyMzQ1Njc4OTAiLCJ0aW1lc3RhbXAiOiIyMDIwLTA3LTEwIDE1OjAwOjAwLjAwMCJ9
          type: string
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: Opaque
      scheme: bearer
      type: http