# openapi/model_account_numbers.go
/*
 * Chariot Payments API
 *
 * The Chariot Payments REST API.
 *
 * API version: v1
 * Contact: developers@givechariot.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// AccountNumbers - The identifying numbers used for making electronic transfers to accounts
type AccountNumbers struct {

	// The ACH numbers
	Ach []AchNumber `json:"ach,omitempty"`
}