# openapi/model_ach_number.go
/*
 * Chariot Payments API
 *
 * The Chariot Payments REST API.
 *
 * API version: v1
 * Contact: developers@givechariot.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

// AchNumber - The identifying number used for making electronic transfers to an account
type AchNumber struct {

	// The account number
	AccountNumber string `json:"account_number,omitempty"`

	// The American Bankers' Association (ABA) Routing Transit Number (RTN).
	RoutingNumber string `json:"routing_number,omitempty"`
}